Do not explicitly depend on latest GKE version
When versions are explicitly set to the latest available version, Annealing almost always fails to apply the patch due to yet-unknown reasons. The rationale for setting the versions explicitly was to ensure that the clusters are always updated in time. But it seems like it is not worth the trouble.

Without the explicit latest versions, the master should still be automatic upgrade (may not be immediate after version availability):

https://cloud.google.com/kubernetes-engine/versioning-and-upgrades#automatic_master_upgrades

We also set "Auto Upgrade" on the nodes, which should upgrades the nodes to master versions (may not be immediate after master version upgrade).

So it seems without these lines, we can still expect the gke versions of the cluster to upgrade (eventually).

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=206408347