Have a resource component receive the "registrar resource" instead of creating it
We have several "consoles" ("views" in the registrar console). They all affect the same resource - the registrar itself.

Currently, each view creates its own "RESTful resource", even though it's the same resource for all of them, meaning we have copied code and copied "URL endpoint" across multiple files.

I assume this was made so that IF one day we have a "view" to another resource, we can easily add it. But we currently don't have any such view, nor plans to add any.

So according to the YAGNI paradigm, it's better to move the resource creation outside of the console.

Also, IF we do add a view to a different resource - it'll still be more readable to have a map from the "view" to the "resource URL endpoint" alongside the existing map from the "view" to the "console"...

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=213992967