Add dependency on diffutils
com_google_truth has a dependency on diffutils that we're not exposing.  This
becomes problematic in cases where certain tests fail and the equality check
can't show the difference.  If this happens, instead of the original failure
all we see is a failure to load diffutils.

Note that com_google_truth appears to have some other dependencies that we are
also not exposing, but that so far these have not been problematic.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=214774587